feature(llm-dispatch): WIP strategy pattern works in debugger
