--- src/llm-dispatch/invoke-chatgpt-API-hello-world-2.py
@@ -1,6 +1,0 @@
-from openai import OpenAI
-
-client = OpenAI()
-completion = client.chat.completions.create(
-    model="gpt-4o", messages=[{"role": "user", "content": "write a haiku about ai"}]
-)
